Reapply "Support pin swap (only PWM channels) in sine firmware also"
This reverts commit 6580a14d2790af09f0ce01cf5ebcf5662f308704.

+++ src/throttle.cpp
@@ -55,7 +55,7 @@ int Throttle::accelmax;
 int Throttle::accelflt;
 float Throttle::maxregentravelhz;
 
-bool Throttle::CheckAndLimitRange(int* potval, int potIdx)
+bool Throttle::CheckAndLimitRange(int* potval, uint8_t potIdx)
 {
    int potMin = potmax[potIdx] > potmin[potIdx] ? potmin[potIdx] : potmax[potIdx];
    int potMax = potmax[potIdx] > potmin[potIdx] ? potmax[potIdx] : potmin[potIdx];
